make Schema serializable
When using BuildSchema it becomes important to cache the output schema.

This can be achieved trivially with something like APC but only if you don’t have closures in your objects and/or you wrap them in SerializableClosures.

I’ve transformed everything so that it’s no serializable and will fit into APC if users are writing their resolvers without closures.