Modifies the chr() check to properly handle \u0000 codes in the range… (#554)
* Modifies the chr() check to properly handle \u0000 codes in the range [127, 255] in the lexer

This caused the lexer to output invalid UTF-8 for input like 'pok\u00E9mon'. The output for the é
would be the decimal byte 233, which would indicate that it should be a 4 byte unicode sequence,
but the following bytes didn't have the leading 10 prefix, so the sequence was invalid UTF-8.