Never resolve additional types in Schema constructor
When you construct a Schema and pass an array of type callables like this:
```php
Schema(['types' => [
  'TypeName' => function() { },
]);
```
the callables are immediately resolved in the constructor.

If you want the types to not be resolved, you must use a callable that returns the types array with callables like this:
```php
Schema(['types' => fn() => [
  'TypeName' => function() { },
]);
```

I was surprised by that because I thought my array of callables was already enough for lazy loading.

When looking at the source code I find the distinction a bit weird and I think it can be handled in the same way.

When `Schema::getTypeMap` is called, it will call `Schema::collectAllTypes` which will call `Schema::resolveAdditionalTypes`.