Fix broken SchemaPrinter
In 914d91e717ac2b5e5dc7b4253726b00185ac745b a breaking change was introduced while
removing `Utils::every`.

Since that change, it only includes descriptions if ALL of the arguments have a description.

If one argument has a description and one has not, no descriptions will be added.

That doesn't make sense.

It should be the opposite.

When none of the arguments have a description, then inline them.