Allow sorting more items in SchemaPrinter (#1292)
This change allows you to configure the SchemaPrinter and define what should be sorted and what not.

`sortArguments` allows to sort arguments used in fields on objects and interfaces.

`sortFields` allows to sort fields inside objects and interfaces.

`sortInputFields` allows to sort fields inside input objects.

`sortEnumValues` allows to sort enum values.

It's important to know that changing the order of fields inside input objects can have severe
consequences for clients that have generated code based on that order. For example, Apollo for
Kotlin and Swift takes the ordering from the input object to create the input objects.

When the order in the schema changes, so does the order of the constructor arguments for those
objects. When the calling code doesn't use named arguments, they can pass wrong values.

That's why I made sorting of input object fields a separate option.