Before fetching a type, first check if it exists
When upgrading from 15.1.0 to 15.2.1 I noticed the following error in our setup:

```
  Type "Subscription" not found in name to fully qualified mapping
```

We have a very strict service locator. Any type that is retrieved should exist, or fail.

This works fine, and allows us to quickly spot invalid referenced types.

But this broke in f82c04f191906150bfb0ca3458f238d24312d839

I think it's better to first check if a type exists, before comparing it.