Replace self:: with static:: in schema printer
The `SchemaPrinter` class uses static methods.

Those methods can be overloaded in a child class (they are protected). However, if we want the base `SchemaPrinter` static function to be able to call the overloaded functions in the child class, the call to these functions must be done using the `static::` keyword (Late static binding).

This is done properly almost everywhere in the code, but in some places there remains calls done using `self::`.

This commit turns the remaining `self::` into `static::` to make Late Static Binding possible.