Allow lazy root types (query, mutation, subscription)
This makes it possible to create a schema with the root types lazy. This way, only when the root
type is needed, the type will be initialized by calling the callable.

Why would a Schema have an instance of the mutation and subscription root type, when it's only
serving a query operation? Visa versa. That does not make sense to me.