Fix scoping of context for mutations
In #1385 we introduced the `ScopedContext` interface. When implemented on the Context, it would clone
the context down to it's children.

This works great for queries. But mutations are executed serially. I forgot to call `maybeScopeContext` there.

This fixes the problem, and adds a test to guard it.it would clone the context down to it's children.