Allow replacing a Node with a NodeList in Visitor (#1754)
With this change, it's now possible to do something like this:

```php
$editedAst = Visitor::visit(
    $ast,
    [
        NodeKind::INLINE_FRAGMENT => function ($node) {
            return $node->selectionSet->selections;
        },
    ]
);
```

It replaces an inline fragment (Node) with the selections (NodeList).

This makes it possible to create specific optimizer Visitors that cleanup operations.

For example, remove useless inline fragments on the same type. Or inline fragments
that don't have a type selection and also don't use directives.